requirements.txt
pillow
torch
torchvision
tqdm
scikit-learn
transformers